# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paper Setting API",
    "description": "Local-first DOCX thesis formatting compiler",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rule-packs/{rule_pack_id}": {
      "get": {
        "tags": [
          "rule-packs"
        ],
        "summary": "Get Rule Pack",
        "operationId": "get_rule_pack_api_v1_rule_packs__rule_pack_id__get",
        "parameters": [
          {
            "name": "rule_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePack"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rule-packs/{rule_pack_id}/capabilities": {
      "get": {
        "tags": [
          "rule-packs"
        ],
        "summary": "Get Rule Pack Capabilities",
        "operationId": "get_rule_pack_capabilities_api_v1_rule_packs__rule_pack_id__capabilities_get",
        "parameters": [
          {
            "name": "rule_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePackCapabilityReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rule-packs/drafts": {
      "post": {
        "tags": [
          "rule-packs"
        ],
        "summary": "Create Rule Pack Draft",
        "operationId": "create_rule_pack_draft_api_v1_rule_packs_drafts_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_rule_pack_draft_api_v1_rule_packs_drafts_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePackDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rule-packs/import": {
      "post": {
        "tags": [
          "rule-packs"
        ],
        "summary": "Import Rule Pack",
        "operationId": "import_rule_pack_api_v1_rule_packs_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_rule_pack_api_v1_rule_packs_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePackSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvancedWordFormat": {
        "properties": {
          "formula": {
            "$ref": "#/components/schemas/FormulaFormat"
          },
          "citations": {
            "$ref": "#/components/schemas/CitationFormat"
          },
          "toc": {
            "$ref": "#/components/schemas/TocFormat"
          },
          "numbering": {
            "$ref": "#/components/schemas/AutomaticNumberingFormat"
          },
          "cross_references": {
            "$ref": "#/components/schemas/CrossReferenceFormat"
          },
          "notes": {
            "$ref": "#/components/schemas/NotesFormat"
          }
        },
        "type": "object",
        "title": "AdvancedWordFormat"
      },
...
      "AutomaticNumberingFormat": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "headings": {
            "type": "boolean",
            "title": "Headings",
            "default": true
          },
          "bibliography": {
            "type": "boolean",
            "title": "Bibliography",
            "default": true
          },
          "captions": {
            "type": "boolean",
            "title": "Captions",
            "default": true
          },
          "strip_existing_prefix": {
            "type": "boolean",
            "title": "Strip Existing Prefix",
            "default": true
          }
        },
        "type": "object",
        "title": "AutomaticNumberingFormat"
      },
...
      "Body_create_rule_pack_draft_api_v1_rule_packs_drafts_post": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Pasted formatting guide text"
          },
          "document": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document",
            "description": "UTF-8 TXT, Markdown, or DOCX formatting guide"
          },
          "rule_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Pack Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "Body_create_rule_pack_draft_api_v1_rule_packs_drafts_post"
      },
...
      "Body_import_rule_pack_api_v1_rule_packs_import_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document",
            "description": "JSON rule pack"
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "Body_import_rule_pack_api_v1_rule_packs_import_post"
      },
...
      "CharacterFormat": {
        "properties": {
          "east_asia_font": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "East Asia Font"
          },
          "latin_font": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latin Font"
          },
          "size_pt": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Pt"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9A-Fa-f]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "bold": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bold"
          },
          "italic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Italic"
          },
          "underline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Underline"
          }
        },
        "type": "object",
        "title": "CharacterFormat"
      },
      "CitationFormat": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "style": {
            "type": "string",
            "const": "numeric_brackets",
            "title": "Style",
            "default": "numeric_brackets"
          },
          "sort_numbers": {
            "type": "boolean",
            "title": "Sort Numbers",
            "default": true
          },
          "collapse_ranges": {
            "type": "boolean",
            "title": "Collapse Ranges",
            "default": true
          },
          "range_separator": {
            "type": "string",
            "enum": [
              "-",
              "–"
            ],
            "title": "Range Separator",
            "default": "–"
          }
        },
        "type": "object",
        "title": "CitationFormat"
      },
...
      "FormulaFormat": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "math_font": {
            "type": "string",
            "minLength": 1,
            "title": "Math Font",
            "default": "Cambria Math"
          },
          "style": {
            "type": "string",
            "enum": [
              "preserve",
              "plain",
              "italic",
              "bold",
              "bold_italic"
            ],
            "title": "Style",
            "default": "preserve"
          }
        },
        "type": "object",
        "title": "FormulaFormat"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageFormat": {
        "properties": {
          "width_mm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Width Mm",
            "default": 210
          },
          "height_mm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Height Mm",
            "default": 297
          },
          "orientation": {
            "type": "string",
            "enum": [
              "portrait",
              "landscape"
            ],
            "title": "Orientation",
            "default": "portrait"
          },
          "margin_top_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Margin Top Mm",
            "default": 25.4
          },
          "margin_bottom_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Margin Bottom Mm",
            "default": 25.4
          },
          "margin_left_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Margin Left Mm",
            "default": 25.4
          },
          "margin_right_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Margin Right Mm",
            "default": 25.4
          },
          "header_distance_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Header Distance Mm",
            "default": 15
          },
          "footer_distance_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Footer Distance Mm",
            "default": 17.5
          }
        },
        "type": "object",
        "title": "PageFormat"
      },
      "ParagraphFormat": {
        "properties": {
          "alignment": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "left",
                  "center",
                  "right",
                  "justify"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment"
          },
          "first_line_indent_pt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Line Indent Pt"
          },
          "left_indent_pt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left Indent Pt"
          },
          "right_indent_pt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Right Indent Pt"
          },
          "space_before_pt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Before Pt"
          },
          "space_after_pt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space After Pt"
          },
          "space_before_lines": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Before Lines"
          },
          "space_after_lines": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Space After Lines"
          },
          "line_spacing": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Spacing"
          },
          "line_spacing_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "multiple",
                  "exact",
                  "at_least"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Spacing Mode"
          },
          "keep_with_next": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keep With Next"
          },
          "keep_together": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keep Together"
          },
          "page_break_before": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Break Before"
          }
        },
        "type": "object",
        "title": "ParagraphFormat"
      },
...
      "RequirementSource": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "builtin",
              "text",
              "docx",
              "template",
              "manual"
            ],
            "title": "Kind",
            "default": "manual"
          },
          "label": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Label"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9A-Fa-f]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "RequirementSource"
      },
...
      "RoleRule": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[a-z0-9_.-]+$",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "title": "Role"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "error"
            ],
            "title": "Severity",
            "default": "error"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 100
          },
          "min_confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Confidence",
            "default": 0.9
          },
          "preserve_style_identity": {
            "type": "boolean",
            "title": "Preserve Style Identity",
            "default": false
          },
          "paragraph": {
            "$ref": "#/components/schemas/ParagraphFormat"
          },
          "character": {
            "$ref": "#/components/schemas/CharacterFormat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "description",
          "source"
        ],
        "title": "RoleRule"
      },
      "RuleEvidence": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "property_path": {
            "type": "string",
            "title": "Property Path"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Value"
          },
          "source_line": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Source Line"
          },
          "quote": {
            "type": "string",
            "maxLength": 240,
            "title": "Quote"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "role",
          "property_path",
          "value",
          "source_line",
          "quote",
          "confidence"
        ],
        "title": "RuleEvidence"
      },
      "RulePack": {
        "properties": {
          "schema_version": {
            "type": "string",
            "enum": [
              "1.0",
              "2.0"
            ],
            "title": "Schema Version",
            "default": "1.0"
          },
          "id": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[a-z0-9-]+$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "version": {
            "type": "string",
            "minLength": 1,
            "title": "Version"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "zh-CN"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "requirements": {
            "$ref": "#/components/schemas/RuleRequirements"
          },
          "page": {
            "$ref": "#/components/schemas/PageFormat"
          },
          "explicit_page_fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit Page Fields"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleRule"
            },
            "type": "array",
            "minItems": 1,
            "title": "Roles"
          },
          "advanced": {
            "$ref": "#/components/schemas/AdvancedWordFormat"
          },
          "template": {
            "$ref": "#/components/schemas/TemplateRule"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "roles"
        ],
        "title": "RulePack"
      },
      "RulePackCapabilityReport": {
        "properties": {
          "executable": {
            "type": "boolean",
            "title": "Executable"
          },
          "required": {
            "items": {
              "type": "string",
              "enum": [
                "document.page_layout",
                "document.paragraph_formatting",
                "document.character_formatting",
                "word.formula_formatting",
                "word.citation_normalization",
                "word.toc",
                "word.automatic_numbering",
                "word.notes",
                "template.inspect",
                "template.fill",
                "template.placeholders",
                "sections.page_numbering",
                "sections.headers_footers",
                "word.cross_references",
                "word.content_controls"
              ]
            },
            "type": "array",
            "title": "Required"
          },
          "supported": {
            "items": {
              "type": "string",
              "enum": [
                "document.page_layout",
                "document.paragraph_formatting",
                "document.character_formatting",
                "word.formula_formatting",
                "word.citation_normalization",
                "word.toc",
                "word.automatic_numbering",
                "word.notes",
                "template.inspect",
                "template.fill",
                "template.placeholders",
                "sections.page_numbering",
                "sections.headers_footers",
                "word.cross_references",
                "word.content_controls"
              ]
            },
            "type": "array",
            "title": "Supported"
          },
          "unsupported": {
            "items": {
              "type": "string",
              "enum": [
                "document.page_layout",
                "document.paragraph_formatting",
                "document.character_formatting",
                "word.formula_formatting",
                "word.citation_normalization",
                "word.toc",
                "word.automatic_numbering",
                "word.notes",
                "template.inspect",
                "template.fill",
                "template.placeholders",
                "sections.page_numbering",
                "sections.headers_footers",
                "word.cross_references",
                "word.content_controls"
              ]
            },
            "type": "array",
            "title": "Unsupported"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers"
          }
        },
        "type": "object",
        "required": [
          "executable",
          "required",
          "supported",
          "unsupported"
        ],
        "title": "RulePackCapabilityReport"
      },
      "RulePackDraft": {
        "properties": {
          "rule_pack": {
            "$ref": "#/components/schemas/RulePack"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/RuleEvidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "source_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Filename"
          },
          "source_format": {
            "type": "string",
            "enum": [
              "text",
              "txt",
              "md",
              "docx",
              "mixed"
            ],
            "title": "Source Format"
          },
          "recognized_properties": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Recognized Properties"
          },
          "unrecognized_lines": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unrecognized Lines"
          },
          "capability_report": {
            "$ref": "#/components/schemas/RulePackCapabilityReport"
          }
        },
        "type": "object",
        "required": [
          "rule_pack",
          "evidence",
          "warnings",
          "source_format",
          "recognized_properties",
          "unrecognized_lines",
          "capability_report"
        ],
        "title": "RulePackDraft"
      },
      "RulePackSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "builtin": {
            "type": "boolean",
            "title": "Builtin"
          },
          "capability_report": {
            "$ref": "#/components/schemas/RulePackCapabilityReport"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "locale",
          "content_hash",
          "builtin",
          "capability_report"
        ],
        "title": "RulePackSummary"
      },
      "RuleRequirements": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/RequirementSource"
            },
            "type": "array",
            "title": "Sources"
          },
          "unresolved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unresolved"
          },
          "conflicts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "required_capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "document.page_layout",
                "document.paragraph_formatting",
                "document.character_formatting",
                "word.formula_formatting",
                "word.citation_normalization",
                "word.toc",
                "word.automatic_numbering",
                "word.notes",
                "template.inspect",
                "template.fill",
                "template.placeholders",
                "sections.page_numbering",
                "sections.headers_footers",
                "word.cross_references",
                "word.content_controls"
              ]
            },
            "type": "array",
            "title": "Required Capabilities"
          }
        },
        "type": "object",
        "title": "RuleRequirements"
      },
...
      "TemplateFillRule": {
        "properties": {
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Values"
          }
        },
        "type": "object",
        "title": "TemplateFillRule"
      },
...
      "TemplateRule": {
        "properties": {
          "fill": {
            "$ref": "#/components/schemas/TemplateFillRule"
          },
          "structure": {
            "$ref": "#/components/schemas/TemplateStructureRule"
          }
        },
        "type": "object",
        "title": "TemplateRule"
      },
...
      "TocFormat": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "min_level": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Min Level",
            "default": 1
          },
          "max_level": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Max Level",
            "default": 3
          },
          "hyperlinks": {
            "type": "boolean",
            "title": "Hyperlinks",
            "default": true
          },
          "update_on_open": {
            "type": "boolean",
            "title": "Update On Open",
            "default": true
          }
        },
        "type": "object",
        "title": "TocFormat"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}